requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
attrs==24.2.0
binance-futures-connector==4.1.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dateparser==1.2.0
Flask==3.0.3
frozenlist==1.5.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.0
packaging==24.2
pandas==2.2.3
plotly==5.24.1
propcache==0.2.1
pycryptodome==3.21.0
python-binance==1.0.25
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
setuptools==75.6.0
six==1.17.0
ta==0.11.0
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.0.6
yarl==1.18.3
zipp==3.21.0
